Relative paths for debug prints
If the user provides a define -DCHK_SRC_DIR_LENGTH,
then the debug prints that use CHK_FILE instead of __FILE
will skip the part of the path before the project's root.
Ofcourse, if you have a path that's not relative to the root,
DON'T use CHK_PATH as that would eat part of the path.